+++ lib/ahoy/tracker.rb
@@ -12,6 +12,7 @@ module Ahoy
       @controller = options[:controller]
       @request = options[:request] || @controller.try(:request)
       @visit_token = options[:visit_token]
+      @user = options[:user]
       @options = options
       @user = options[:user]
     end
@@ -59,7 +60,7 @@ module Ahoy
 
           @store.track_visit(data)
 
-          Ahoy::GeocodeV2Job.perform_later(visit_token, data[:ip]) if Ahoy.geocode
+          Ahoy::GeocodeV2Job.perform_later(visit_token, data[:ip]) if Ahoy.geocode && data[:ip]
         end
       end
       true
@@ -130,7 +131,7 @@ module Ahoy
     end
 
     def visit_properties
-      @visit_properties ||= request.present? ? Ahoy::VisitProperties.new(request, api: api?).generate : {}
+      @visit_properties ||= request ? Ahoy::VisitProperties.new(request, api: api?).generate : {}
     end
 
     def visit_token
@@ -170,7 +171,7 @@ module Ahoy
 
     def set_cookie(name, value, duration = nil, use_domain = true)
       # safety net
-      return if request.nil? || !Ahoy.cookies
+      return unless Ahoy.cookies && request
 
       cookie = {
         value: value
@@ -182,9 +183,7 @@ module Ahoy
     end
 
     def delete_cookie(name)
-      return if request.nil?
-
-      request.cookie_jar.delete(name) if request.cookie_jar[name] }
+      request.cookie_jar.delete(name) if request && request.cookie_jar[name]
     end
 
     def trusted_time(time = nil)
@@ -200,8 +199,12 @@ module Ahoy
     end
 
     def report_exception(e)
-      raise e if !defined?(Rails) || Rails.env.development? || Rails.env.test?
-      Safely.report_exception(e)
+      if defined?(ActionDispatch::RemoteIp::IpSpoofAttackError) && e.is_a?(ActionDispatch::RemoteIp::IpSpoofAttackError)
+        debug "Tracking excluded due to IP spoofing"
+      else
+        raise e if !defined?(Rails) || Rails.env.development? || Rails.env.test?
+        Safely.report_exception(e)
+      end
     end
 
     def generate_id